+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `apiRequireFaction` returns also user avatar
 - User submitted news have player avatar as originator
 - Faction resources changed from `Int` to `RawResource a`
+- Faction name changed from `Text` to `FactionName`
 ### Removed
 - `maybeFaction` function removed
 ### Fixed